Update github-actions-ec2.yml
Using cp instead of mv makes sure that there is a copy of files present. The -- flag tells the mv command to treat the next argument as the destination directory, rather than a source file. The !(README.md) expression is a negation pattern that matches all files except README.md